Sequre history delete and some changes

+++ src/Provider/ItemProvider.jsx
@@ -1,8 +1,11 @@
 import React, { createContext, useEffect, useState } from 'react';
+import { toast } from 'react-toastify';
 import Swal from 'sweetalert2';
 export const ItemContext = createContext(null);
 const ItemProvider = ({ children }) => {
+    const notify = () => toast("You don't have access to delete")
     const currentDate = new Date();
+    const accessKey = import.meta.env.VITE_accessKey;
     const options = { year: 'numeric', month: 'long', day: 'numeric', hour: 'numeric', minute: 'numeric', hour12: true };
     const formattedDate = currentDate.toLocaleString('en-US', options);
     const [items, setItems] = useState([]);
@@ -14,34 +17,40 @@ const ItemProvider = ({ children }) => {
     }, [items])
     //Handle Delete
     const handleDelete = (key) => {
-        Swal.fire({
+        const userKey = prompt('Please provide access key');
+        if (userKey == accessKey) {
+            Swal.fire({
 
-            title: 'Are you sure?',
-            text: "You won't be able to revert this!",
-            icon: 'warning',
-            showCancelButton: true,
-            confirmButtonColor: '#3085d6',
-            cancelButtonColor: '#d33',
-            confirmButtonText: 'Yes, delete it!',
-        }).then((result) => {
-            if (result.isConfirmed) {
-                fetch(`http://localhost:5000/${key}`, {
-                    method: 'DELETE',
-                    headers: { 'Content-type': 'application/json' },
-                })
-                    .then(res => res.json())
-                    .then(data => {
-
-
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!',
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    fetch(`http://localhost:5000/${key}`, {
+                        method: 'DELETE',
+                        headers: { 'Content-type': 'application/json' },
                     })
-                Swal.fire(
-                    'Deleted!',
-                    'Your file has been deleted.',
-                    'success'
-                )
-            }
-
-        })
+                        .then(res => res.json())
+                        .then(data => {
+    
+    
+                        })
+                    Swal.fire(
+                        'Deleted!',
+                        'Your file has been deleted.',
+                        'success'
+                    )
+                }
+    
+            })
+        } else {
+            notify();
+        }
+     
     }
     const orderItem = items.filter(item => item.quantity < 5);
     const itemInfo = {
